refactor: consolidate all k8s apis under top level api/
