+++ repo-agent/cmd/configdir-controller/main.go
@@ -31,7 +31,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	configdirv1alpha1 "github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/configdir/api/v1alpha1"
+	configdirv1alpha1 "github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/api/configdir/v1alpha1"
 	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/pkg/controllers/configdir"
 	//+kubebuilder:scaffold:imports
 )
