+++ repo-agent/scripts/ops/setup-syncer-creds.sh
